Rename mock helper
